Moved CRUD types to their own file
